Split into multiple files and support variables
