+++ packages/serverless-api/src/api.ts
@@ -1,0 +1,11 @@
+export {
+  activateBuild,
+  triggerBuild,
+  waitForSuccessfulBuild,
+} from './internals/builds';
+export { getDependencies } from './internals/dependencies';
+export {
+  getOrCreateFunctionResources,
+  uploadFunction,
+} from './internals/functions';
+export { setEnvironmentVariables } from './internals/variables';
