+++ packages/serverless-api/src/utils/fs.ts
