+++ packages/serverless-api/src/utils/sleep.ts
@@ -1,0 +1,2 @@
+export const sleep = (ms: number) =>
+  new Promise(resolve => setTimeout(resolve, ms));
